requirements.txt
pypdf==4.0.1
python-docx==0.8.11
pdfplumber==0.10.3
Flask==3.0.0
werkzeug==3.0.0
openai==1.54.3
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
jsonschema==4.19.2
nltk==3.8.1
spacy==3.7.2
pathlib2==2.3.7
regex==2023.10.3
pytest==7.4.3
black==23.10.1
flake8==6.1.0
gunicorn==22.0.0